python-cli/pyproject.toml
[project]
name = "python-cli"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["go-authgate==0.2.2", "python-dotenv>=1.0.0"]
